Basically re-wrote the entire site
Was running into a bunch of issues with unfortunately starting the site from a desktop view, so I started over and went mobile first like I should have. The site came together much easier and is now mobile compliant.

Also removed the dropdown menus, because ultimately they werent needed for the final product of the site.